src/test/java/com/codeknight/recipeapp/controllers/RecipeControllerTest.java
package com.codeknight.recipeapp.controllers;

import com.codeknight.recipeapp.domain.Recipe;
import com.codeknight.recipeapp.services.RecipeService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MockMvcBuilder;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.ui.Model;

import static org.junit.jupiter.api.Assertions.*;

class RecipeControllerTest {
    private RecipeController controller;

    @Mock
    private RecipeService recipeService;

    private MockMvc mockMvc;

    @Mock
    private Model model;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.initMocks(this);
        controller = new RecipeController(recipeService);
        mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
    }

    @Test
    void testGetRecipeByIdController() throws Exception{
        mockMvc.perform(MockMvcRequestBuilders.get("/recipe/show/1"))
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.view().name("recipe/show"));
    }

    @Test
    void getRecipeById() {
        String recipeId = "1";
        String viewModel = controller.getRecipeById(recipeId, model);

        assertEquals("recipe/show", viewModel);
    }


}
